Splits point test into submodules, replaces test globals with module fixtures.
