+++ tests/test_point/test_point_arithmetic.py
@@ -1,7 +1,6 @@
 import contextlib
 
 import pytest
-from cryptography.exceptions import InternalError
 from cryptography.hazmat.backends.openssl import backend
 from cryptography.hazmat.primitives.asymmetric import ec
 
@@ -10,13 +9,10 @@ from umbral.config import default_curve
 from umbral.point import Point
 
 curve = default_curve() or ec.SECP256K1()
-RANDOM_K256_POINT = Point.gen_rand(curve)
-RANDOM_K256_POINT2 = Point.gen_rand(curve)
-RANDOM_K256_BIGNUM = BigNum.gen_rand(curve)
 
 
 @pytest.fixture()
-def mock_openssl(mocker):
+def mock_openssl(mocker, random_ec_point1, random_ec_bignum):
     @contextlib.contextmanager
     def mocked_openssl_backend():
         def mocked_ec_point_equality(group, ec_point, other_point, bn_ctx):
@@ -25,9 +21,9 @@ def mock_openssl(mocker):
             assert 'EC_POINT' in str(ec_point)
             assert 'EC_POINT' in str(other_point)
 
-            assert RANDOM_K256_POINT.group == group
-            assert RANDOM_K256_POINT.ec_point == ec_point
-            assert RANDOM_K256_POINT.ec_point == other_point
+            assert random_ec_point1.group == group
+            assert random_ec_point1.ec_point == ec_point
+            assert random_ec_point1.ec_point == other_point
 
             return 1
 
@@ -36,8 +32,8 @@ def mock_openssl(mocker):
             assert 'EC_POINT' in str(other_point)
             assert 'EC_POINT' in str(sum)
 
-            assert RANDOM_K256_POINT.group == group
-            assert RANDOM_K256_POINT.ec_point == ec_point
+            assert random_ec_point1.group == group
+            assert random_ec_point1.ec_point == ec_point
 
             return 1
 
@@ -48,11 +44,11 @@ def mock_openssl(mocker):
             assert 'EC_POINT' in str(product)
             assert 'NULL' in str(null)
 
-            assert RANDOM_K256_POINT.group == group
-            assert RANDOM_K256_POINT.ec_point == ec_point
+            assert random_ec_point1.group == group
+            assert random_ec_point1.ec_point == ec_point
 
-            assert RANDOM_K256_BIGNUM.bignum == bignum
-            assert RANDOM_K256_BIGNUM.curve_nid == RANDOM_K256_POINT.curve_nid
+            assert random_ec_bignum.bignum == bignum
+            assert random_ec_bignum.curve_nid == random_ec_point1.curve_nid
 
             return 1
 
@@ -60,7 +56,7 @@ def mock_openssl(mocker):
             assert 'EC_POINT' in str(inverse)
             assert 'BN_CTX' in str(bn_ctx)
 
-            assert RANDOM_K256_POINT.group == group
+            assert random_ec_point1.group == group
 
             return 1
 
@@ -73,83 +69,16 @@ def mock_openssl(mocker):
             for method, patch in mock_load.items():
                 stack.enter_context(mocker.mock_module.patch.object(backend._lib, method, patch))
             yield
+
     return mocked_openssl_backend
 
 
-def test_mocked_point_curve_arithmetic(mock_openssl):
+def test_mocked_openssl_point_arithmetic(mock_openssl, random_ec_point1, random_ec_point2, random_ec_bignum):
     with mock_openssl():
-        _ = RANDOM_K256_POINT == RANDOM_K256_POINT     # __eq__
-        _ = RANDOM_K256_POINT * RANDOM_K256_BIGNUM     # __mul__
-        _ = RANDOM_K256_POINT + RANDOM_K256_POINT2     # __add__
-        _ = ~RANDOM_K256_POINT                         # __invert__
-
-
-def test_from_to_bytes():
-
-    p = Point.gen_rand(curve)
-    pbytes = p.to_bytes(is_compressed=False)
-    q = Point.from_bytes(pbytes, curve)
-    assert p == q
-
-    pbytes = p.to_bytes(is_compressed=True)
-    q = Point.from_bytes(pbytes, curve)
-    assert p == q
-
-
-def test_point_to_cryptography_pubkey():
-    p = Point.gen_rand(curve)
-
-    crypto_pub_key = p.to_cryptography_pub_key()
-
-    p_affine = p.to_affine()
-    crypto_affine = (
-        crypto_pub_key.public_numbers().x,
-        crypto_pub_key.public_numbers().y
-    )
-
-    assert p_affine == crypto_affine
-
-
-def test_invalid_points():
-    p = Point.gen_rand(curve)
-
-    pbytes = bytearray(p.to_bytes(is_compressed=False))
-    # Flips last bit
-    pbytes[-1] = pbytes[-1] ^ 0x01
-    pbytes = bytes(pbytes)
-
-    try:
-        q = Point.from_bytes(pbytes, curve)
-    except InternalError as e:
-        # We want to catch specific InternalExceptions: 
-        # - Point not in the curve (code 107)
-        # - Invalid compressed point (code 110)
-        # https://github.com/openssl/openssl/blob/master/include/openssl/ecerr.h#L228
-        if e.err_code[0].reason in (107, 110):
-            pass
-        else:
-            assert False
-    else:
-        assert False
-
-
-def test_generator_point():
-    g1 = Point.get_generator_from_curve(curve)
-
-    # http://www.secg.org/SEC2-Ver-1.0.pdf
-    # Section 2.7.1
-    g_compressed = 0x0279BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798
-    g_compressed = g_compressed.to_bytes(32+1, byteorder='big')
-
-    g_uncompressed = 0x0479BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8
-    g_uncompressed = g_uncompressed.to_bytes(64+1, byteorder='big')
-
-    g2 = Point.from_bytes(g_compressed, curve)
-    assert g1 == g2
-
-    g3 = Point.from_bytes(g_uncompressed, curve)
-    assert g1 == g3
-    assert g2 == g3
+        _ = random_ec_point1 == random_ec_point1   # __eq__
+        _ = random_ec_point1 * random_ec_bignum    # __mul__
+        _ = random_ec_point1 + random_ec_point2    # __add__
+        _ = ~random_ec_point1                      # __invert__
 
 
 def test_point_curve_mult_regression():
@@ -175,5 +104,3 @@ def test_point_curve_mult_regression():
     # Repeating the operation, should return the same result.
     product_with_star_operator_again = k256_point * k256_bn
     assert product_with_star_operator == product_with_star_operator_again
-
-
